refactored websocket server
