+++ ws-server/static/index.html
@@ -22,7 +22,9 @@
 		}
 
 		ws.onmessage = function(event) {
-			console.log(event.data);
+			// console.log(event.data);
+			const obj = JSON.parse(event.data);
+			console.log(obj[0]['Timestamp']);
 			title.innerHTML = event.data;
 		}
